backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
pandas
beautifulsoup4
PyMuPDF
python-pptx
requests
faiss-cpu
tiktoken
